Update .dark-My-Posh.omp.json
A misunderstanding was found in the advanced theme. Opening sub levels for bash(wsl) would read a admin terminal when it was just another sub level terminal. This is now resolved, to use admin permission on a linux (ubuntu) terminal use the sudo -s command and apply the theme to root